+++ app/src/main/java/com/wtb/dashTracker/ui/fragment_expenses/ExpenseListFragment.kt
@@ -80,7 +80,7 @@ class ExpenseListFragment : ListItemFragment() {
             ConfirmType.DELETE.key
         ) { _, bundle ->
             val result = bundle.getBoolean(ARG_CONFIRM)
-            val id = bundle.getInt(ARG_EXTRA)
+            val id = bundle.getLong(ARG_EXTRA)
             if (result) {
                 viewModel.deleteExpenseById(id)
             }
